render: Multithreading support
Make a basic thread pool for rendering pixels. Each thread grabs a
pixel, renders its sub-rays (if applicable), writes the resulting color
back to the framebuffer, and repeats.

Could've done it by the ray instead of by the pixel, but by the pixel is
easier to write and debug.

Also added a nice little progress bar, thanks stack overflow