render: Remove rays buffer, generate rays on the fly
Pre-generating all of the rays for the scene is pretty pointless, but it
also means that the program takes forever to initialize (it has to do a
ton of memory allocations) and can't do a 4k render on my system because
it runs out of both RAM and pagefile. This is smarter.